[CCUBE-2142][IT] Rename file and move typing to source file
