+++ src/modal-v2/slots/card.tsx
@@ -2,7 +2,7 @@ import clsx from "clsx";
 import type { NamedExoticComponent } from "react";
 import React from "react";
 
-import isStyledComponent from "../../util/isStyledComponents";
+import isStyledComponent from "../../util/is-styled-components";
 import type { ModalCardProps } from "../types";
 import { CloseButton } from "./close-button";
 import { Content } from "./content";
