+++ src/util/utility-types.ts
@@ -32,5 +32,3 @@ export type KeyOf<T extends object> = Extract<keyof T, string>;
 
 // Get the value types of an object as a union type
 export type ValueOf<T> = T[keyof T];
-
-export type StyledComponentBrand = { readonly _sc: symbol };
